pythonwithdeps/requirements.txt
cycler==0.11.0
distlib==0.3.5
docutils==0.19
et-xmlfile==1.1.0
filelock==3.7.1
Kivy==2.1.0
Kivy-Garden==0.1.5
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.3
openpyxl==3.0.10
pandas==1.4.3
Pillow==8.4.0
platformdirs==2.5.2
polars==0.13.50
psutil==5.9.4
Pygments==2.13.0
pyparsing==3.0.4
python-dateutil==2.8.2
pytz==2022.1
Shapely==1.8.0
tk==0.1.0
typing_extensions==4.2.0
odoorpc
python-vlc
